third_party/blink/web_tests/svg/animations/animate-text-nested-transforms-expected.txt

SVG 1.1 dynamic animation tests
This is some text.

test to determine whether auto-rotate animateMotion path animations pre-muliply or post-multiply animation transform matrix

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rootSVGElement.getBBox().x is 117
PASS rootSVGElement.getBBox().y is -338
PASS rootSVGElement.getBBox().x is 367
PASS rootSVGElement.getBBox().y is 550
PASS successfullyParsed is true

TEST COMPLETE
